Add vue router and add navbar
Refactored the app structure to make it work with vue router
