+++ package.json
@@ -17,11 +17,13 @@
     "chart.js": "^3.7.0",
     "core-js": "^3.6.5",
     "vue": "^3.0.0",
-    "vue-chart-3": "^3.0.9"
+    "vue-chart-3": "^3.0.9",
+    "vue-router": "^4.0.0-0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
     "@vue/compiler-sfc": "^3.0.0",
     "babel-eslint": "^10.1.0",
