+++ src/main.js
@@ -3,4 +3,6 @@ import { Chart, registerables } from 'chart.js'
 Chart.register(...registerables)
 
 import App from './App.vue'
-createApp(App).mount('#app')
+import router from './router'
+
+createApp(App).use(router).mount('#app')
